add the splash screen back in!
